Abstracted mesos-style.py to wrap the cpp linter in a class.
Previously, mesos-style.py was just a collection of functions that
checked the style of relevant files in the mesos code base.  However,
the script assumed that we always wanted to run cpplint over every
file we were checking. Since we are planning on adding a python linter
to the codebase soon, it makes sense to abstract the common
functionality from this script into a class so that a cpp-based linter
and a python-based linter can inherit the same set of common
functionality.

This commit builds this abstraction and implements a 'CppLinter()' in
terms of it.

Review: https://reviews.apache.org/r/50907